app/requirements.txt
flask
cryptography
pycryptodome
firebase_admin